미술, 한우 카테고리 수정 12

+++ src/main/java/com/moaguide/service/art/ArtService.java
@@ -15,6 +15,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.util.List;
@@ -54,7 +55,7 @@ public class ArtService {
                 (String) result[2],  // platform
                 (String) result[3],  // name
                 ((Long) result[4]),  // recruitmentPrice (Integer 그대로 사용)
-                ((Double) result[5]),  // recruitmentRate (Double 그대로 사용)
+                ((BigDecimal) result[5]).doubleValue(),  // recruitmentRate (Double 그대로 사용)
                 ((Long) result[6]),  // totalPrice
                 String.valueOf(result[7]),  // subscriptionDate
                 ((Integer) result[8])  // minInvestment (Integer 그대로 사용)
